reafactor : User와 Auth 패키지 구조 분리
- 깨지는 User쪽 테스트 수정
