+++ src/main/java/bit/auth/dto/AccessTokenCreateRequest.java
@@ -1,6 +1,5 @@
-package bit.user.dto;
+package bit.auth.dto;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
