+++ src/main/java/bit/auth/repository/OAuth2AuthorizationRequestBasedOnCookieRepository.java
@@ -1,4 +1,4 @@
-package bit.user.repository;
+package bit.auth.repository;
 
 import bit.util.CookieUtil;
 import jakarta.servlet.http.Cookie;
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
 import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.stereotype.Component;
 import org.springframework.web.util.WebUtils;
 
 public class OAuth2AuthorizationRequestBasedOnCookieRepository implements
