+++ src/main/java/bit/user/controller/UserController.java
@@ -1,12 +1,8 @@
 package bit.user.controller;
 
-import bit.user.dto.AccessTokenCreateRequest;
-import bit.user.dto.AccessTokenResponse;
 import bit.user.dto.UserResponse;
-import bit.user.service.TokenService;
 import bit.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,22 +12,11 @@ import org.springframework.web.bind.annotation.*;
 public class UserController implements UserControllerDoc {
 
     private final UserService userService;
-    private final TokenService tokenService;
 
     @GetMapping("/email/{email}")
     public ResponseEntity<UserResponse> getUserByEmail(@PathVariable String email) {
         return ResponseEntity.ok(UserResponse.from(userService.getByEmail(email)));
     }
 
-    @PostMapping("/token")
-    public ResponseEntity<AccessTokenResponse> createAccessToken(@RequestBody AccessTokenCreateRequest request) {
-        String newAccessToken = tokenService.createNewAccessToken(request.getRefreshToken());
-        return ResponseEntity.status(HttpStatus.CREATED).body(new AccessTokenResponse(newAccessToken));
-    }
 
-    @PostMapping("/fake/token")
-    public ResponseEntity<AccessTokenResponse> createFakeAccessToken() {
-        String newAccessToken = tokenService.createNewFakeToken();
-        return ResponseEntity.status(HttpStatus.CREATED).body(new AccessTokenResponse(newAccessToken));
-    }
 }
