+++ src/main/java/bit/user/enums/OauthPlatformType.java
@@ -1,4 +1,4 @@
-package bit.user.oauth.enums;
+package bit.user.enums;
 
 public enum OauthPlatformType {
     KAKAO, NAVER, GOOGLE;
