+++ src/test/java/bit/anniversary/service/AnServiceTest.java
@@ -12,7 +12,7 @@ import bit.anniversary.entity.Anniversary;
 import bit.anniversary.repository.AnRepository;
 import bit.user.domain.User;
 import bit.user.entity.UserEntity;
-import bit.user.oauth.enums.OauthPlatformType;
+import bit.user.enums.OauthPlatformType;
 import bit.user.repository.UserJpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
